requirements.txt
requests
brother-ql
pillow
python-fmrest
customtkinter